requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
click==8.3.0
deprecation==2.1.0
Flask==3.1.2
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
protobuf==4.25.8
pydantic==2.11.10
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.31.0
six==1.17.0
tinkoff==0.1.1
tinkoff-investments==0.2.0b116
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
Werkzeug==3.1.3
